requirements.txt
flask
mysql-connector-python
groq
python-dotenv